requirements.txt
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1
kagglehub>=0.2
prophet>=1.1
statsmodels>=0.14
matplotlib>=3.7
pytest>=7.0
